mcp_server/requirements.txt
# MCP Server Dependencies
mcp[cli]==1.9.1
pyairtable>=2.0.0
openai>=1.81.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn>=0.34.0 